Refactor NewBuilder logic, fragmenting logic in to more easily understandable/testable functions and in to the new builder.go file
