+++ pkg/vproj/projects.go
@@ -10,11 +10,9 @@ import (
 	"os"
 	"path/filepath"
 	"strings"
-	"time"
 
 	"github.com/djherbis/buffer"
 	"github.com/djherbis/nio"
-	"github.com/gobwas/glob"
 	"github.com/vorteil/vorteil/pkg/vcfg"
 	"github.com/vorteil/vorteil/pkg/vio"
 	"github.com/vorteil/vorteil/pkg/vpkg"
@@ -166,185 +164,6 @@ func (t *Target) VCFG() (*vcfg.VCFG, error) {
 	}
 
 	return cfg, nil
-}
-
-// NewBuilder ...
-func (t *Target) NewBuilder() (vpkg.Builder, error) {
-
-	b := vpkg.NewBuilder()
-
-	vcfg, err := t.VCFG()
-	if err != nil {
-		return nil, err
-	}
-
-	x, err := vcfg.Marshal()
-	if err != nil {
-		return nil, err
-	}
-
-	cfg := vio.CustomFile(vio.CustomFileArgs{
-		Name:       "default.vcfg",
-		ReadCloser: ioutil.NopCloser(bytes.NewReader(x)),
-		Size:       len(x),
-		ModTime:    time.Now(),
-	})
-
-	err = b.SetVCFG(cfg)
-	if err != nil {
-		return nil, err
-	}
-
-	iconPath := t.Icon
-	if iconPath != "" {
-		if !filepath.IsAbs(iconPath) {
-			iconPath = filepath.Join(t.Dir, t.Icon)
-		}
-
-		var icon vio.File
-		if _, err := os.Stat(iconPath); err != nil {
-			// icon not provided
-			icon = vio.CustomFile(vio.CustomFileArgs{
-				Name:       "default.png",
-				ReadCloser: ioutil.NopCloser(bytes.NewReader([]byte{})),
-				Size:       0,
-				ModTime:    time.Now(),
-			})
-		} else {
-			icon, err = vio.LazyOpen(iconPath)
-			if err != nil {
-				return nil, err
-			}
-
-			if icon.IsDir() {
-				err = fmt.Errorf("icon file is a directory: %s", t.Icon)
-				return nil, err
-			}
-		}
-
-		err = b.SetIcon(icon)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	ignore := make([]glob.Glob, 0)
-	for _, p := range t.Ignore {
-		ip, err := glob.Compile(p)
-		if err != nil {
-			return nil, err
-		}
-
-		ignore = append(ignore, ip)
-	}
-
-	t.Dir, err = filepath.Abs(t.Dir)
-	if err != nil {
-		return nil, err
-	}
-	t.Dir = filepath.ToSlash(t.Dir)
-
-	err = filepath.Walk(t.Dir, func(path string, info os.FileInfo, err error) error {
-
-		if err != nil {
-			return err
-		}
-
-		path = filepath.ToSlash(path)
-		abs := path
-		path = strings.TrimPrefix(path, t.Dir)
-		path = strings.TrimPrefix(path, "/")
-		if path == "" {
-			return nil
-		}
-
-		for _, ip := range ignore {
-			if ip.Match(path) {
-				if info.IsDir() {
-					return filepath.SkipDir
-				}
-				return nil
-			}
-		}
-
-		f, err := vio.LazyOpen(abs)
-		if err != nil {
-			return err
-		}
-
-		err = b.AddToFS(path, f)
-		if err != nil {
-			return err
-		}
-
-		return nil
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	for _, d := range t.Files {
-		abs := d
-		if !filepath.IsAbs(d) {
-			abs = filepath.Join(t.Dir, d)
-		}
-
-		abs = filepath.ToSlash(abs)
-
-		err = filepath.Walk(abs, func(path string, info os.FileInfo, err error) error {
-
-			if err != nil {
-				return err
-			}
-
-			thisAbs := path
-			path = filepath.ToSlash(path)
-			path = strings.TrimPrefix(path, t.Dir)
-			path = strings.TrimPrefix(path, "/")
-
-			f, err := vio.LazyOpen(thisAbs)
-			if err != nil {
-				return err
-			}
-
-			if f.IsDir() {
-				err = filepath.Walk(abs, func(path string, info os.FileInfo, err error) error {
-
-					thisDir := path
-					path = strings.TrimPrefix(path, abs)
-					path = strings.TrimPrefix(path, "/")
-					if path == "" {
-						return nil
-					}
-
-					f, err := vio.LazyOpen(thisDir)
-					if err != nil {
-						return err
-					}
-
-					err = b.AddToFS(path, f)
-					if err != nil {
-						return err
-					}
-
-					return nil
-				})
-			} else {
-				err = b.AddToFS(path, f)
-				if err != nil {
-					return err
-				}
-
-			}
-
-			return nil
-		})
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	return b, nil
 }
 
 func vcfgInfo(cfg *vcfg.VCFG) (vio.File, error) {
